requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Templates and forms
jinja2==3.1.2
python-multipart==0.0.6

# AI and environment
openai==1.106.1
python-dotenv==1.0.0

# Database and authentication
sqlalchemy==2.0.23
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# Health check dependency
requests==2.31.0

# Additional dependencies for production
starlette==0.27.0
pydantic==2.5.0
